:recycle: @Valid MasterExceptionHandler 에서 처리 1. ResponseEntityExceptionHandler extends 를 받지 않음 - 왜냐하면 그럼 @Valid 로 잡히는 것들을 처리를 제대로 못함 - 그래서 직접 MethodArgumentNotValidException 로 ExceptionHandler 에 맵핑하려고 하는데, 저 extends 때문에 중복이라고 Bean 오류 - ResponseEntityExceptionHandler extends 받지 않는 것으로 해결 2. error 반환 시에 result 에 오류 목록들 받아서 처리 하도록 변경 - BaseResponseDto.onFailure 에 T errors 매개변수 추가 - 이걸로 @Valid 상황에도 BaseResponseDto 로 처리 가능 3. MasterExceptionHandler 에서 handle 방식 변경 - handleExceptionInternal - handleExceptionInternalFalse - 위의 함수 모두 body 에 BaseResponseDto 로 만든 오류 메시지 반환하도록
