--- src/main/java/com/umc/ttg/global/common/ErrorResponse.java
@@ -1,40 +1,0 @@
-package com.umc.ttg.global.common;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Getter
-public class ErrorResponse {
-
-    private String message;
-    private String code;
-    private int status;
-    private List<FieldError> errors = new ArrayList<>();
-
-    @Builder
-    public ErrorResponse(String message, String code, int status, List<FieldError> errors) {
-        this.message = message;
-        this.code = code;
-        this.status = status;
-        this.errors = errors;
-    }
-
-    private List<FieldError> initErrors(List<FieldError> errors) {
-        return (errors == null) ? new ArrayList<>() : errors;
-    }
-
-    @Getter
-    public static class FieldError {
-        private String field;
-        private String message;
-
-        @Builder
-        public FieldError(String field, String message) {
-            this.field = field;
-            this.message = message;
-        }
-    }
-}
