+++ src/main/java/com/umc/ttg/global/error/handler/MasterExceptionHandler.java
@@ -1,7 +1,6 @@
 package com.umc.ttg.global.error.handler;
 
 import com.umc.ttg.global.common.BaseResponseDto;
-import com.umc.ttg.global.common.ErrorResponse;
 import com.umc.ttg.global.common.ResponseCode;
 import com.umc.ttg.global.error.GeneralException;
 import lombok.extern.slf4j.Slf4j;
@@ -13,7 +12,6 @@ import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
@@ -42,41 +40,30 @@ public class MasterExceptionHandler {
     }
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ResponseEntity<Object> processValidationError(MethodArgumentNotValidException exception, WebRequest request) {
-        final List<ErrorResponse.FieldError> fieldErrors = getFieldErrors(exception.getBindingResult());
+    public ResponseEntity<Object> processValidationError(MethodArgumentNotValidException exception) {
+
+        final List<BaseResponseDto.FieldError> fieldErrors = getFieldErrors(exception.getBindingResult());
         return ResponseEntity.status(HttpStatus.BAD_REQUEST)
                 .body(buildFieldErrors(ResponseCode._BAD_REQUEST, fieldErrors));
+
     }
 
-//    protected ResponseEntity<Object> handleExceptionInternal(Exception ex, Object body,
-//                                                             HttpHeaders headers, HttpStatus status, WebRequest request) {
-//
-//        log.info("At exception handler");
-//        ServletRequestAttributes requestAttributes = (ServletRequestAttributes) request;
-//        HttpServletRequest servletRequest = requestAttributes.getRequest();
-//
-//        String contentType = request.getHeader("Content-Type");
-//        log.info("Content-Type : {}", contentType);
-//        log.error("발생한 에러의 로그 :", ex);
-//        return handleExceptionInternal(ex, ResponseCode.valueOf(status), headers, status, request);
-//    }
-
-    private List<ErrorResponse.FieldError> getFieldErrors(BindingResult bindingResult) {
+    private List<BaseResponseDto.FieldError> getFieldErrors(BindingResult bindingResult) {
         final List<FieldError> errors = bindingResult.getFieldErrors();
         return errors.parallelStream()
-                .map(error -> ErrorResponse.FieldError.builder()
+                .map(error -> BaseResponseDto.FieldError.builder()
                         .field(error.getField())
                         .message(error.getDefaultMessage()).build())
                 .collect(Collectors.toList());
     }
 
-    private ErrorResponse buildFieldErrors(ResponseCode responseCode, List<ErrorResponse.FieldError> errors) {
-        return ErrorResponse.builder()
+    private BaseResponseDto buildFieldErrors(ResponseCode responseCode, List<BaseResponseDto.FieldError> errors) {
+        return BaseResponseDto.builder()
+                .isSuccess(false)
                 .code(responseCode.getCode())
-                .status(responseCode.getHttpStatus().value())
                 .message(responseCode.getMessage())
-                .errors(errors).build();
+                .result(errors)
+                .build();
     }
 
     private ResponseEntity<Object> handleExceptionInternal(Exception e, ResponseCode errorCode,
@@ -88,15 +75,16 @@ public class MasterExceptionHandler {
 
     private ResponseEntity<Object> handleExceptionInternal(Exception e, ResponseCode errorCode,
                                                            HttpHeaders headers, HttpStatus status, WebRequest request) {
-        BaseResponseDto<Object> body = BaseResponseDto.onFailure(errorCode);
+        BaseResponseDto<Object> body = BaseResponseDto.onFailure(null, errorCode);
         return ResponseEntity.status(errorCode.getHttpStatus())
                 .body(body);
     }
 
     private ResponseEntity<Object> handleExceptionInternalFalse(Exception e, ResponseCode errorCode,
                                                                 HttpHeaders headers, HttpStatus status, WebRequest request) {
-        BaseResponseDto<Object> body = BaseResponseDto.onFailure(errorCode);
+        BaseResponseDto<Object> body = BaseResponseDto.onFailure(null, errorCode);
         return ResponseEntity.status(errorCode.getHttpStatus())
                 .body(body);
     }
+
 }
